university_chatbot/requirements.txt
# =========================================================
# Campus Bot - Django Chatbot Project
# Environment: chatbot_env_312
# Python Version: 3.12 ✅
# NOTE: Do NOT use Python 3.13 (not supported by scikit-learn & some libs yet)
# =========================================================

# ----------------------------
# Django Framework & REST API
# ----------------------------
django==5.0.7
djangorestframework==3.15.2
django-modeltranslation==0.18.11

# ----------------------------
# Database & Utilities
# ----------------------------
pytz==2024.1
sqlparse==0.5.1

# ----------------------------
# Language Detection & Translation
# ----------------------------
googletrans==4.0.0-rc1
langdetect==1.0.9

# ----------------------------
# Document Handling
# ----------------------------
PyPDF2==3.0.1
python-pptx==0.6.23

# ----------------------------
# Machine Learning / NLP Utilities
# ----------------------------
scikit-learn==1.3.2
numpy==1.26.4

# ----------------------------
# Asynchronous I/O & Typing
# ----------------------------
sniffio==1.3.0
trio==0.25.0
outcome==1.2.0
exceptiongroup==1.2.0
typing_extensions==4.12.0

# ----------------------------
# Testing
# ----------------------------
pytest==8.2.0

# ----------------------------
# Optional (uncomment if needed)
# ----------------------------
# uvloop==0.19.0   # High-performance asyncio event loop (Linux/macOS only)